refactor: isolate profile onboarding protocols and routes
Changed:
Separated legacy onboarding persistence, learner routes, operator preview routes, and typed run request parsing from their shared modules.

Benefit:
Rolling compatibility and modern profile flows remain unchanged while their ownership and transaction boundaries are easier to review.
